Move option parsing to separate file
